pyproject.toml
[project]
name = "my-langgraph-gemini-agent"
version = "0.1.0"
dependencies = [
    "langgraph",
    "langchain-google-genai",
    "langserve",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "sse-starlette",
    "gradio", 
]